requirements.txt
streamlit
pandas
bcrypt
plotly
openai
fpdf
streamlit-extras
google-generativeai
